Move cookie method to JEEContext
