+++ pac4j-core/src/main/java/org/pac4j/core/context/JEEContext.java
@@ -6,6 +6,10 @@ import org.pac4j.core.util.CommonHelper;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.*;
 
 /**
@@ -15,6 +19,61 @@ import java.util.*;
  * @since 1.4.0
  */
 public class JEEContext implements WebContext {
+
+    private static ZoneId GMT = ZoneId.of("GMT");
+
+    /**
+     * Date formats with time zone as specified in the HTTP RFC to use for formatting.
+     * @see <a href="https://tools.ietf.org/html/rfc7231#section-7.1.1.1">Section 7.1.1.1 of RFC 7231</a>
+     */
+    private static DateTimeFormatter DATE_FORMATTER = DateTimeFormatter.ofPattern("EEE, dd MMM yyyy HH:mm:ss zzz", Locale.US).withZone(GMT);
+
+    /**
+     * Custom method for adding cookie because the servlet-api version doesn't support SameSite attributes.
+     * @param cookie pac4j Cookie object
+     * @param response Http Response
+     */
+    private static void addCookieHeaderToResponse(Cookie cookie, final HttpServletResponse response) {
+        var builder = new StringBuilder();
+        builder.append(String.format("%s=%s;", cookie.getName(), cookie.getValue()));
+
+        if (cookie.getMaxAge() > -1) {
+            builder.append(String.format(" Max-Age=%s;", cookie.getMaxAge()));
+            long millis = cookie.getMaxAge() > 0 ? System.currentTimeMillis() + (cookie.getMaxAge() * 1000) : 0;
+            Instant instant = Instant.ofEpochMilli(millis);
+            ZonedDateTime time = ZonedDateTime.ofInstant(instant, GMT);
+            builder.append(String.format(" Expires=%s;", DATE_FORMATTER.format(time)));
+        }
+        if (CommonHelper.isNotBlank(cookie.getDomain())) {
+            builder.append(String.format(" Domain=%s;", cookie.getDomain()));
+        }
+        builder.append(String.format(" Path=%s;", CommonHelper.isNotBlank(cookie.getPath()) ? cookie.getPath() : "/"));
+
+        var sameSitePolicy = cookie.getSameSitePolicy() == null ? "none" : cookie.getSameSitePolicy().toLowerCase();
+        switch (sameSitePolicy) {
+            case "strict":
+                builder.append(" SameSite=Strict;");
+                break;
+            case "lax":
+                builder.append(" SameSite=Lax;");
+                break;
+            case "none":
+            default:
+                builder.append(" SameSite=None;");
+                break;
+        }
+        if (cookie.isSecure() || "none".equals(sameSitePolicy)) {
+            builder.append(" Secure;");
+        }
+        if (cookie.isHttpOnly()) {
+            builder.append(" HttpOnly;");
+        }
+        var value = builder.toString();
+        if (value.endsWith(";")) {
+            value = value.substring(0, value.length() - 1);
+        }
+        response.addHeader("Set-Cookie", value);
+    }
 
     private final HttpServletRequest request;
 
@@ -150,8 +209,6 @@ public class JEEContext implements WebContext {
         if (cookies != null) {
             for (final var c : cookies) {
                 final var cookie = new Cookie(c.getName(), c.getValue());
-                cookie.setVersion(c.getVersion());
-                cookie.setComment(c.getComment());
                 cookie.setDomain(c.getDomain());
                 cookie.setHttpOnly(c.isHttpOnly());
                 cookie.setMaxAge(c.getMaxAge());
@@ -164,8 +221,8 @@ public class JEEContext implements WebContext {
     }
 
     @Override
-    public void addResponseCookie(final Cookie cookie) {
-        Cookie.addCookieHeaderToResponse(cookie, response);
+    public void addResponseCookie(Cookie cookie) {
+        addCookieHeaderToResponse(cookie, this.response);
     }
 
     /**
